completed openmp code and provided slurm files
Forgot to say that the main code specifies 20 threads rn. That won't be a problem for dawson or I because our programs max out at 20 threads and they will only use 1 thread, even if you specify 20, if only one node is available. So nothing will need to be changed as dawson and I test different cores. Nick however might need to set his threads to 40